Generate blank placeholders for virtual Texture2D
The Asset Dumper doesn't support dumping virtual textures, so it skips
creating a '.png' file in the asset dump. However, the Asset Generator
does not take this into account and happily crashes when trying to use
the missing image file.

To avoid having to manually delete any entries for virtual textures in
the asset dump, read the value of the 'VirtualTextureStreaming' member
of the asset's JSON descriptor and, if true, generate a blank image as
placeholder. In addition, log a warning to easily identify assets that
could not be generated.

TEST=Dump and generate assets for Satisfactory, successfully.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>