chore: Drop `EngineVersion` field
The `EngineVersion` field cannot be parsed and causes a warning. Given
that the version specified there must exactly match that of the engine
version in use (which is a custom version), setting the correct engine
version in there means it has to be changed every time the engine gets
updated. There's no benefit in using this for Satisfactory modding, so
just drop the field.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>